Create meeting.html
Barebones HTML for the Meeting info and minutes page with navigation links. "Meeting Info" section is optional and kept minimal for privacy purposes. I am not sure how we would like to set up this page to provide the minutes - do we want to set it up so that viewers can download documentation of the minutes as a file? A link to a Google Doc? Do we want a viewable document embedded in the page itself?